requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/dev.in
#
about-time==3.1.1
    # via alive-progress
alive-progress==2.1.0
    # via -r requirements/dev.in
antlr4-python3-runtime==4.8
    # via
    #   hydra-core
    #   omegaconf
anyio==3.5.0
    # via
    #   httpcore
    #   watchgod
asgiref==3.5.0
    # via uvicorn
attrs==21.4.0
    # via pytest
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.0.12
    # via
    #   httpx
    #   requests
click==8.1.2
    # via uvicorn
commonmark==0.9.1
    # via rich
coverage[toml]==6.3.2
    # via pytest-cov
fastapi==0.68.1
    # via -r requirements/dev.in
grapheme==0.6.0
    # via alive-progress
gunicorn==20.1.0
    # via -r requirements/dev.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.14.7
    # via httpx
httptools==0.2.0
    # via uvicorn
httpx==0.22.0
    # via -r requirements/dev.in
hydra-core==1.1.1
    # via -r requirements/dev.in
idna==3.3
    # via
    #   anyio
    #   requests
    #   rfc3986
importlib-resources==5.4.0
    # via hydra-core
iniconfig==1.1.1
    # via pytest
omegaconf==2.1.1
    # via hydra-core
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pydantic==1.9.0
    # via fastapi
pygments==2.11.2
    # via rich
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via
    #   -r requirements/dev.in
    #   pytest-cov
pytest-cov==3.0.0
    # via -r requirements/dev.in
python-dotenv==0.20.0
    # via uvicorn
pyyaml==6.0
    # via
    #   omegaconf
    #   uvicorn
requests==2.27.1
    # via -r requirements/dev.in
rfc3986[idna2008]==1.5.0
    # via httpx
rich==12.0.1
    # via -r requirements/dev.in
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.14.2
    # via fastapi
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.1.1
    # via pydantic
urllib3==1.26.9
    # via requests
uvicorn[standard]==0.15.0
    # via -r requirements/dev.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.8.2
    # via uvicorn
websockets==10.2
    # via uvicorn
zipp==3.6.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
